task finished

+++ src/main/java/com/codecool/marsexploration/data/SimulationInput.java
@@ -1,4 +1,4 @@
 package com.codecool.marsexploration.data;
 
-public record SimulationInput(String mapPath, Coordinate landing, long timeout, String logPath) {
+public record SimulationInput(String mapPath, Coordinate landing, int timeout, String logPath) {
 }
